+++ Code/Mantid/Framework/SINQ/inc/MantidSINQ/PoldiUtilities/IPoldiFunction1D.h
@@ -49,7 +49,7 @@ public:
   IPoldiFunction1D();
   virtual ~IPoldiFunction1D() {}
 
-  virtual void poldiFunction1D(const std::vector<int> &indices,
+  virtual void poldiFunction1D(const std::vector<size_t> &indices,
                                const API::FunctionDomain1D &domain,
                                API::FunctionValues &values) const = 0;
 };
