+++ Code/Mantid/Framework/SINQ/inc/MantidSINQ/PoldiUtilities/Poldi2DFunction.h
@@ -54,7 +54,7 @@ public:
   virtual void function1DSpectrum(const API::FunctionDomain1DSpectrum &domain,
                                   API::FunctionValues &values) const;
 
-  virtual void poldiFunction1D(const std::vector<int> &indices,
+  virtual void poldiFunction1D(const std::vector<size_t> &indices,
                                const API::FunctionDomain1D &domain,
                                API::FunctionValues &values) const;
