+++ Code/Mantid/Framework/SINQ/inc/MantidSINQ/PoldiUtilities/PoldiSpectrumDomainFunction.h
@@ -250,7 +250,7 @@ public:
   functionDeriv1DSpectrum(const API::FunctionDomain1DSpectrum &domain,
                           API::Jacobian &jacobian);
 
-  void poldiFunction1D(const std::vector<int> &indices,
+  void poldiFunction1D(const std::vector<size_t> &indices,
                        const API::FunctionDomain1D &domain,
                        API::FunctionValues &values) const;
