+++ Code/Mantid/Framework/SINQ/inc/MantidSINQ/PoldiUtilities/PoldiSpectrumPawleyFunction.h
@@ -47,7 +47,7 @@ public:
                           API::FunctionValues &values) const;
   void functionDeriv1DSpectrum(const API::FunctionDomain1DSpectrum &domain,
                                API::Jacobian &jacobian);
-  void poldiFunction1D(const std::vector<int> &indices,
+  void poldiFunction1D(const std::vector<size_t> &indices,
                        const API::FunctionDomain1D &domain,
                        API::FunctionValues &values) const;
